Removed the step to use a form to add an arrangement, now it's a fast process
