+++ src/components/navAppBar/navAppBar.jsx
@@ -8,7 +8,10 @@ import { withStyles } from '@material-ui/core/styles';
 import { AppBar, Toolbar, Typography } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
 import { setAccount, logout } from 'actions/account/account'
+import { setRealData } from 'actions/real/real'
+import { uuid } from 'utils'
 import config from 'config.json'
+
 
 const styles = {
     root: {
@@ -35,6 +38,33 @@ class NavAppBar extends Component {
 
     goToViewAllArrangements = () => {
         this.props.history.push("/all_arrangements")
+    }
+
+    createNewArrangement = () => {
+        var d = new Date()
+        const arrangement_id = uuid("arrangement")
+        let real = {
+            user: this.props.account.user.googleId,
+            _id: arrangement_id,
+            containers: [],
+            is_deleted: false,
+            items: [],
+            name: "Untitled Arrangement",
+            users: [this.props.account.user.googleId],
+            owner: this.props.account.user.googleId,
+            snapshots: [{
+                _id: uuid("snapshot"),
+                name: "Version 1",
+                snapshot: {}, // TO REMOVE after it is removed on backend
+                snapshotContainers: [],
+                unassigned: []
+            }],
+            timestamp: d.getTime() / 1000,
+            modified_timestamp: d.getTime() / 1000
+        }
+
+        this.props.setRealData(real)
+        this.props.history.push('/arrangement/' + arrangement_id)
     }
 
     onFailure = (error) => {
@@ -71,7 +101,7 @@ class NavAppBar extends Component {
         if (this.props.account.isAuthenticated) {
             buttons = (
                 <div>
-                    <Button onClick={this.goToNewArrangement} color="inherit">
+                    <Button onClick={this.createNewArrangement} color="inherit">
                         New
                     </Button>
                     <Button onClick={this.goToViewAllArrangements} color="inherit">
@@ -132,6 +162,9 @@ const mapDispatchToProps = (dispatch, ownProps) => {
         },
         setAccount: (account) => {
             dispatch(setAccount(account))
+        },
+        setRealData: (data) => {
+            dispatch(setRealData(data))
         }
     }
 }
