+++ src/config/router.jsx
@@ -8,7 +8,6 @@ import createHistory from 'history/createBrowserHistory'
 import Layout from 'layout'
 import HomePage from 'views/homePage/homePage'
 import Arrangement from 'views/arrangement/arrangement'
-import AddArrangement from 'views/addArrangement/addArrangement'
 import AllArrangements from 'views/allArrangements/allArrangements'
 
 const history = createHistory()
@@ -19,7 +18,6 @@ const router = (
             <Switch>
                 <Route exact path='/' component={HomePage} />
                 <PrivateRoute exact path='/arrangement/:arrangement_id' component={Arrangement} />
-                <PrivateRoute exact path='/addarrangement' component={AddArrangement} />
                 <PrivateRoute exact path='/all_arrangements' component={AllArrangements} />
             </Switch>
         </Layout>
